Strip out sqlalchemy-migrate.
This loses schema management, which was one of the main aims of mortar_rdb, but sqlalchemy-migrate was painful. Maybe come back with alembic?

In the meantime, mortar_rdb makes a good way to manage mutli-session projects and multi-package projects that all supply models. The helpers for database creations and cleansing, along with testing, are still useful.